fix(chat): make file/image paths clickable in new chat thread
Extract the clickable file-link chips out of AgentMessage into a shared
fileLinkChips module and wire them into ChatMarkdown, so generated file
paths (e.g. /tmp/pr-card.png) open an inline preview in the split panel.

Generated-By: PostHog Code
Task-Id: 27dba876-5444-406d-888f-e58ec6368782
